+++ lib/graphql/subscriptions/subscription_root.rb
@@ -42,7 +42,7 @@ module GraphQL
             # for the backend to register:
             event = Subscriptions::Event.new(
               name: field.name,
-              arguments: arguments,
+              arguments: arguments_without_field_extras(arguments: arguments),
               context: context,
               field: field,
             )
@@ -50,7 +50,7 @@ module GraphQL
             value
           elsif context.query.subscription_topic == Subscriptions::Event.serialize(
               field.name,
-              arguments,
+              arguments_without_field_extras(arguments: arguments),
               field,
               scope: (field.subscription_scope ? context[field.subscription_scope] : nil),
             )
@@ -60,6 +60,14 @@ module GraphQL
           else
             # This is a subscription update, but this event wasn't triggered.
             context.skip
+          end
+        end
+
+        private
+
+        def arguments_without_field_extras(arguments:)
+          arguments.dup.tap do |event_args|
+            field.extras.each { |k| event_args.delete(k) }
           end
         end
       end